Add files via upload
I have done a mini project as a part of the s5 academic project under the KTU syllabus. Here our project was on "WEB APPLICATION FOR ONLINE DRIVER AUTHENTICATION SYSTEM" and we have done its design. The main idea under this topic is that the designing of a web application where RTO acts as an administrator and the user can upload documents as well as complaints where the police officials can retrieve from the app and provide notifications as required.